ui/requirements.txt
Flask==3.0.3
Flask_Mail==0.9.1
matplotlib==3.7.5
numpy==1.24.4
Pillow==10.3.0
psycopg2_binary==2.9.9
tqdm==4.66.1
